requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.3
python-magic==0.4.27
